Implement OCR content setting and title generation actions
